Create toggle functionality to display city search results
Added logic to toggle the display of city search results depending on
the number of results returned, enhancing user experience. Before this
change, the search results were always visible, which was not optimal
when the search returned only one city. This will improve the user
interface of the city search feature. No related issues.